+++ CHANGELOG.md
@@ -4,8 +4,14 @@
 
 - added help texts explaining the asterisk in forms (#305)
 - added reference to analyses' download files which user has
-  uploaded a topography file (#188) 
-
+  uploaded a topography file (#188)
+- unified "please wait" indicators when loading analysis results by AJAX (#236)
+- showing notification time stamps now in local time zone
+- showing now 403 page if entering a URL which tries to
+  access a surface which is no longer accesible
+- showing now 404 page also for topography urls which do not exist
+  (before server error) 
+ 
 ## 0.7.0
 
 - surfaces and topographies can now be tagged (#52)